+++ src/clj/rems/api/workflows.clj
@@ -15,14 +15,16 @@
    (s/optional-key :forms) [{:form/id s/Int}]
    :type (apply s/enum events/workflow-types) ; TODO: exclude master workflow?
    (s/optional-key :handlers) [schema-base/UserId]
-   (s/optional-key :licenses) [schema-base/LicenseId]})
+   (s/optional-key :licenses) [schema-base/LicenseId]
+   (s/optional-key :disable-commands) [schema-base/DisableCommandRule]})
 
 (s/defschema EditWorkflowCommand
   {:id s/Int
    (s/optional-key :organization) schema-base/OrganizationId
    ;; type can't change
    (s/optional-key :title) s/Str
-   (s/optional-key :handlers) [schema-base/UserId]})
+   (s/optional-key :handlers) [schema-base/UserId]
+   (s/optional-key :disable-commands) [schema-base/DisableCommandRule]})
 
 (s/defschema CreateWorkflowResponse
   {:success s/Bool